[Fix] impl all function in setup.py to avoid failed build
